feat: 에피그램 리스트 컴포넌트로 에피그램 컴포넌트 감싸는 형식으로 수정
